+++ features/step_definitions/timeout.rb
@@ -1,3 +1,0 @@
-When /^the client is idle for (\S+) seconds$/ do |seconds|
-  sleep seconds.to_f
-end