+++ Changelog.md
@@ -8,12 +8,14 @@ API Changes
 * Added {Ftpd::FtpServer#max_connections}
 * Added {Ftpd::FtpServer#max_connections_per_ip}
 * Added {Ftpd::FtpServer#max_failed_logins}
+* Added {Ftpd::FtpServer#failed_login_delay}
 
 Enhancements
 
 * Configurable maximum connections (defaults to 200).
 * Configurable maximum connections per IP (defaults to 5).
 * Disconnect after too many failed login attempts (configurable).
+* Delay after failed login (configurable).
 
 ### 0.5.0